% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_type.R
\name{auth_type}
\alias{auth_type}
\alias{query_auth_type}
\alias{bearer_auth_type}
\alias{header_auth_type}
\title{Auth types}
\usage{
auth_type(type, ...)

query_auth_type(param_names)

bearer_auth_type()

header_auth_type(header)
}
\arguments{
\item{type}{the authorization type}

\item{...}{other values to attach to the authorization type.}

\item{param_names}{the parameter names to include in the query}

\item{header}{the name of the header where the credential goes}
}
\value{
a wrapify_auth object
}
\description{
Functions for declaring different kinds of authorization for a wrapper.
}
\section{Functions}{
\itemize{
\item \code{query_auth_type()}: Authorization in the request query

\item \code{bearer_auth_type()}: Bearer token

\item \code{header_auth_type()}: authorization in header

}}
