% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapify.R
\name{credential_setter}
\alias{credential_setter}
\title{Create a credential manager function for an API wrapper}
\usage{
credential_setter(wrapper)
}
\arguments{
\item{wrapper}{The API wrapper object initialized using the \code{wrapper} function.}
}
\value{
A custom credential manager function that stores the API credentials
in the environment variable specified in the wrapper. When called,
the returned function collects the API credentials from the user
using \code{ask_for_credentials} and stores them as a JSON string
in the environment variable.
}
\description{
This function generates a custom credential manager function for the given
API wrapper. The credential manager function is responsible for storing the
API credentials in the environment variable specified in the wrapper.
This function only works when the \code{key_management} option of the wrapper
is set to "environment".
}
