% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_type.R
\name{format.wrapify_auth}
\alias{format.wrapify_auth}
\title{Format a wrapify auth object}
\usage{
\method{format}{wrapify_auth}(x, ...)
}
\arguments{
\item{x}{A wrapify_auth object}

\item{...}{Additional arguments (unused)}
}
\value{
A character vector representing the auth type
}
\description{
Format a wrapify auth object
}
