% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapify.R
\name{format.wrapify_wrapper}
\alias{format.wrapify_wrapper}
\title{Format a wrapify wrapper object}
\usage{
\method{format}{wrapify_wrapper}(x, ...)
}
\arguments{
\item{x}{A wrapify_wrapper object}

\item{...}{Additional arguments (unused)}
}
\value{
A character vector representing the wrapper
}
\description{
Format a wrapify wrapper object
}
