% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapify.R
\name{function_args}
\alias{function_args}
\title{Function arguments}
\usage{
function_args(...)
}
\arguments{
\item{...}{named function arguments. Add a value to supply a default, otherwise leave blank.
See example.}
}
\description{
This is a simple wrapper for creating a pairlist, the data structure used for representing
function args. Use this to create arguments for a requestor function.
}
\examples{
simple_wrapper <- wrapper("http://abc.com")
args <- function_args(id = , type = "text") # `id` will have no default, `type` will have default
# value "text"
simple_requestor <- requestor(
  wrapper = simple_wrapper,
  resource = "information/{type}/{id}",
  resource_args = args,
  perform_by_default = FALSE
)
simple_requestor(id = 123)$url
simple_requestor(id = 123, "hello")$url

}
