% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{generate_roxygen_comment}
\alias{generate_roxygen_comment}
\title{Generate a Roxygen comment}
\usage{
generate_roxygen_comment(
  func,
  title = deparse(substitute(func)),
  description = "[Add a description here]",
  param_descriptions = list(),
  return_description = "[Describe the return value here]",
  examples = FALSE,
  export = TRUE,
  cat = TRUE
)
}
\arguments{
\item{func}{The function to document}

\item{title}{The title of the function documentation (default: deparse(substitute(func)))}

\item{description}{The description of the function being documented (default: "\link{Add a description here}")}

\item{param_descriptions}{Description of the parameters of the function, in a named list (default: list())}

\item{return_description}{Description of the return value (default: "\link{Describe the return value here}")}

\item{examples}{Do you want to include examples? (default: FALSE)}

\item{export}{Do you want the function to be exported? (default: TRUE)}

\item{cat}{if TRUE, this will print the output on screen and return invisibly (default: TRUE)}
}
\value{
A string containing the generated Roxygen comment
}
\description{
Generates a roxygen comment for a function for documentation.
}
