% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_type.R
\name{print.wrapify_auth}
\alias{print.wrapify_auth}
\title{Print a wrapify auth object}
\usage{
\method{print}{wrapify_auth}(x, ...)
}
\arguments{
\item{x}{A wrapify_auth object}

\item{...}{Additional arguments passed to format}
}
\value{
Invisibly returns the original object
}
\description{
Print a wrapify auth object
}
