% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapify.R
\name{print.wrapify_wrapper}
\alias{print.wrapify_wrapper}
\title{Print a wrapify wrapper object}
\usage{
\method{print}{wrapify_wrapper}(x, ...)
}
\arguments{
\item{x}{A wrapify_wrapper object}

\item{...}{Additional arguments passed to format}
}
\value{
Invisibly returns the original object
}
\description{
Print a wrapify wrapper object
}
