% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapify.R
\name{requestor}
\alias{requestor}
\title{Create an API requestor function}
\usage{
requestor(
  wrapper,
  resource,
  resource_args = NULL,
  resource_constants = NULL,
  query_args = NULL,
  query_constants = NULL,
  body_args = NULL,
  body_constants = NULL,
  header_args = NULL,
  header_constants = NULL,
  method = "get",
  content_type = NULL,
  perform_by_default = TRUE,
  extract_body_by_default = TRUE,
  extractor = "infer"
)
}
\arguments{
\item{wrapper}{The wrapper object initialized using the \code{wrapper} function.}

\item{resource}{The API endpoint to be called (e.g., "/users/\{user_id\}").}

\item{resource_args}{A named list of arguments to be used as path parameters in the returned function. Create using \code{function_args()}.}

\item{resource_constants}{A named list of constants to be used as path parameters.}

\item{query_args}{A named list of arguments to be used as query parameters in the returned function. Create using \code{function_args()}.}

\item{query_constants}{A named list of constants to be used as query parameters.}

\item{body_args}{A named list of arguments to be used as request body parameters in the returned function. Create using \code{function_args()}.}

\item{body_constants}{A named list of constants to be used as request body parameters.}

\item{header_args}{A named list generated by \code{function_args} of arguments that will be passed to the headers.}

\item{header_constants}{A named list of constants to appear in headers.}

\item{method}{The HTTP method to use for the API request (default: "get"). Accepts "get", "post", "put", "delete", etc.}

\item{content_type}{The content type for the API request (default: determined by the wrapper's default_content_type).}

\item{perform_by_default}{should the returned function try to perform the request by default?}

\item{extract_body_by_default}{If true, the value will be extracted from the response and returned.
Otherwise, the response is returned.}

\item{extractor}{Either "infer", which chooses an extractor based on the content type,
or a function that takes an httr2 response and returns the data that you want.}
}
\value{
A custom requestor function for making API calls with the specified settings. The returned function will have the specified resource_args, query_args, and other arguments.
}
\description{
This function generates a custom requestor function for the given wrapper
and API endpoint. The generated function can be used to make API calls with
specified parameters and configurations. The resource_args, query_args, and
other argument lists will be the arguments of the returned function. These
should be created using the \code{function_args} function.
}
