% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleconstructor.R
\name{super_simple_constructor}
\alias{super_simple_constructor}
\title{Super Simple Constructor}
\usage{
super_simple_constructor(
  ...,
  .class = c("supersimpleconstruct", "list"),
  .prune = TRUE
)
}
\arguments{
\item{...}{The arguments for the constructor function.}

\item{.class}{A character string representing the custom class for the constructor function's output (default: 'supersimpleconstruct').}

\item{.prune}{A logical value determining whether NULL values should be removed from the output named list (default: TRUE).}
}
\value{
A constructor function that takes the specified arguments and returns a named list with an associated class.
}
\description{
Generates a constructor function with the specified arguments, an optional custom class, and
a prune option. The constructor function returns a named list containing the values of the arguments.
}
\examples{
# Create a constructor function with 'a' and 'b' as arguments
my_constructor <- super_simple_constructor(a = , b = , .class = "my_custom_class")

# Call the constructor function with values for 'a' and 'b'
result <- my_constructor(a = 1, b = 2)

# Check the class of the 'result' variable
class(result) # Output: "my_custom_class"
}
