% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapify.R
\name{requestor}
\alias{requestor}
\title{Create an API requestor function}
\usage{
requestor(
  wrapper,
  resource,
  resource_args = NULL,
  resource_constants = NULL,
  query_args = NULL,
  query_constants = NULL,
  body_args = NULL,
  body_constants = NULL,
  method = "get",
  content_type = NULL,
  body_type = "json",
  additional_request_args = NULL,
  perform_by_default = TRUE
)
}
\arguments{
\item{wrapper}{The wrapper object initialized using the \code{wrapper} function.}

\item{resource}{The API endpoint to be called (e.g., "/users/{user_id}").}

\item{resource_args}{A named list of arguments to be used as path parameters in the returned function. Create using \code{function_args()}.}

\item{resource_constants}{A named list of constants to be used as path parameters.}

\item{query_args}{A named list of arguments to be used as query parameters in the returned function. Create using \code{function_args()}.}

\item{query_constants}{A named list of constants to be used as query parameters.}

\item{body_args}{A named list of arguments to be used as request body parameters in the returned function. Create using \code{function_args()}.}

\item{body_constants}{A named list of constants to be used as request body parameters.}

\item{method}{The HTTP method to use for the API request (default: "get"). Accepts "get", "post", "put", "delete", etc.}

\item{content_type}{The content type for the API request (default: determined by the wrapper's default_content_type).}

\item{body_type}{The type of request body to use for the API request (default: 'json').}

\item{additional_request_args}{Additional arguments to be passed to the request object.}

\item{default_action}{The default action to perform when the generated requestor function is called (default: "perform"). Accepts "perform" or "dryrun".}

\item{decode_if_success_default_value}{A logical value indicating whether to decode the response if the request is successful (default: TRUE).}

\item{post_process_default_value}{A logical value indicating whether to apply the post_processor by default.}

\item{post_processor}{A function that is applied to the return value. You might want to use as_tibble or something.
If NULL then the return value is returned unchanged.}
}
\value{
A custom requestor function for making API calls with the specified settings. The returned function will have the specified resource_args, query_args, and other arguments.
}
\description{
This function generates a custom requestor function for the given wrapper
and API endpoint. The generated function can be used to make API calls with
specified parameters and configurations. The resource_args, query_args, and
other argument lists will be the arguments of the returned function. These
should be created using the \code{function_args} function.
}
