% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapify-package.R
\docType{package}
\name{wrapify-package}
\alias{wrapify}
\alias{wrapify-package}
\title{wrapify: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Colin Fraser \email{colin.r.fraser@gmail.com}

}
\keyword{internal}
