% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapify-package.R
\docType{package}
\name{wrapify-package}
\alias{wrapify}
\alias{wrapify-package}
\title{wrapify: Quickly create API wrappers}
\description{
Declaratively produce API wrappers in a few lines of code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/colin-fraser/wrapify}
  \item Report bugs at \url{https://github.com/colin-fraser/wrapify/issues}
}

}
\author{
\strong{Maintainer}: Colin Fraser \email{colin.r.fraser@gmail.com}

}
\keyword{internal}
