% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapify.R
\name{wrapper}
\alias{wrapper}
\title{Initialize a wrapper for an API}
\usage{
wrapper(
  base_url,
  default_headers = list(`content-type` = "application/json"),
  default_query_args = list(),
  auth_type = "none",
  ...
)
}
\arguments{
\item{base_url}{The base URL}

\item{default_headers}{The default headers for API requests}

\item{default_query_args}{A named list of default query arguments to include in all API requests.}

\item{auth_type}{The authentication type}

\item{...}{Additional arguments passed to the wrapper object.}
}
\value{
A wrapper object with the provided settings.
}
\description{
This function creates a wrapper object for an API with the given settings.
The created wrapper can be used with the requestor function to make API calls.
}
